Generate AST class members (ParserContext, GetChildNodes)
The generated AST now contains the `ParserContext` property (a reference to the ANTLR parse tree context) and a `GetChildNodes` method to get the immediate child nodes of an AST node.

Both can be used by the (LSP) language server - to find the node at a specific position in the document, etc.