Use special Missing nodes if parse fails
Since any syntax node reference can be null when parsing incomplete input (especially when writing in the editor), we would have to mark all properties in the AST as nullable (optional) and all code using these would have check for null.

Instead, we fill the "holes" with special nodes that have `IsMissing == true` and have default values of all properties.