Decouple Generators from Configuration
Necessary refactoring of how the configuration values flow to the Generators. Now we're able to specify default/fallback values that can depend on other values, mark which properties are required, do post-processing on the config values into some other form, etc.

The Generators now expose all of their inputs as init-only properties that are mapped from Configuration in the `FromConfig` methods.